+++ docs/demos/app.py
@@ -132,6 +132,5 @@ lhl = Lihil(
 lhl.static("/ping", "pong")
 
 
-breakpoint()
 # if __name__ == "__main__":
 #     lhl.run(__file__)